% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_survey.R
\docType{data}
\name{coffee_survey}
\alias{coffee_survey}
\title{The Great American Coffee Taste Test data}
\format{
A data frame with 4,042 rows and 6 columns:
\describe{
\item{prefer_overall}{ Respondent's preferred coffee type}
\item{gender}{Respondent's gender}
\item{age}{Respondent's age group}
\item{where_drink}{Typical location where coffee is consumed}
\item{brew}{Coffee brewing method used}
\item{most_paid}{Respondent's age group}
...
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/blob/master/data/2024/2024-05-14/readme.md}
}
\usage{
coffee_survey
}
\description{
The data comes from the "Great American Coffee Taste Test" conducted by James Hoffmann and Cometeer in October 2023.
Survey ...
}
\keyword{datasets}
