% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_func.R
\name{launch_app}
\alias{launch_app}
\title{The Shiny App}
\usage{
launch_app()
}
\value{
It will open a new browser window displaying the Shiny app.
}
\description{
This shiny app is for Analysis of The Great American Coffee Taste Test.
}
\examples{
\dontrun{
launch_app()
}
}
