% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_func.R
\name{missing_data_summary}
\alias{missing_data_summary}
\title{Missing Data Summary}
\usage{
missing_data_summary(data)
}
\arguments{
\item{data}{A data frame containing the coffee survey data.}
}
\value{
A data frame with the number and percentage of missing values for each column.
}
\description{
This function provides a summary of missing values in the coffee survey dataset.
}
\examples{
\dontrun{
missing_data_summary(coffee_survey)
}
}
